Remove commented code and cast angle value
Removal of old commented code. Added a cast to int32 to the angle value of the movement mesh model to ensure negative values are supported.

+++ subsys/bluetooth/mesh/vnd/movement_srv.c
@@ -15,7 +15,7 @@ struct bt_mesh_movement_set extract_movement(struct net_buf_simple *buf)
 {
 	struct bt_mesh_movement_set mov_conf;
     mov_conf.time = net_buf_simple_pull_be32(buf);
-    mov_conf.angle = net_buf_simple_pull_be32(buf);
+    mov_conf.angle = (int32_t)net_buf_simple_pull_be32(buf);
     mov_conf.speed = net_buf_simple_pull_u8(buf);
 	return mov_conf;
 }
@@ -36,7 +36,6 @@ static int handle_message_movement_set(struct bt_mesh_model *model, struct bt_me
 
 	BT_MESH_MODEL_BUF_DEFINE(ack, BT_MESH_MOVEMENT_OP_MOVEMENT_ACK, 0);
     bt_mesh_model_msg_init(&ack, BT_MESH_MOVEMENT_OP_MOVEMENT_ACK);
-    // net_buf_simple_add_u8(&ack, 0);
     err = bt_mesh_model_send(model, ctx, &ack, NULL, NULL);
     if (err)
     {
